replaced float by double in class Color

+++ src/Color.h
@@ -40,25 +40,25 @@ class Color
 		Color () : _rgb(0) {}
 		Color (UInt32 rgb) : _rgb(rgb)          {}
 		Color (UInt8 r, UInt8 g, UInt8 b)       {set(r,g,b);}
-		Color (float r, float g, float b)       {set(r,g,b);}
-		Color (const std::vector<float> &rgb)   {set(rgb[0], rgb[1], rgb[2]);}
+		Color (double r, double g, double b)       {set(r,g,b);}
+		Color (const std::vector<double> &rgb)   {set(rgb[0], rgb[1], rgb[2]);}
 		Color (const char *name);
 		Color (const std::string &name);
 		operator UInt32 () const                {return _rgb;}
 		bool operator == (const Color &c) const {return _rgb == c._rgb;}
 		bool operator != (const Color &c) const {return _rgb != c._rgb;}
 		void set (UInt8 r, UInt8 g, UInt8 b)    {_rgb = (r << 16) | (g << 8) | b;}
-		void set (float r, float g, float b);
+		void set (double r, double g, double b);
 		bool set (std::string name, bool case_sensitive=true);
 		void setGray (UInt8 g)                  {set(g,g,g);}
-		void setGray (float g)                  {set(g,g,g);}
-		void setHSB (float h, float s, float b);
-		void setCMYK (float c, float m, float y, float k);
-		void getRGB (float &r, float &g, float &b) const;
+		void setGray (double g)                  {set(g,g,g);}
+		void setHSB (double h, double s, double b);
+		void setCMYK (double c, double m, double y, double k);
+		void getRGB (double &r, double &g, double &b) const;
 		void operator *= (double c);
 		std::string rgbString () const;
-		static void CMYK2RGB (const std::vector<float> &cmyk, std::vector<float> &rgb);
-		static void HSB2RGB (const std::vector<float> &hsb, std::vector<float> &rgb);
+		static void CMYK2RGB (const std::vector<double> &cmyk, std::vector<double> &rgb);
+		static void HSB2RGB (const std::vector<double> &hsb, std::vector<double> &rgb);
 
 	private:
 		UInt32 _rgb;
